requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
arrow==1.3.0
async-timeout==5.0.1
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
blis==0.7.11
cachelib==0.9.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
confection==0.1.5
cookiecutter==2.6.0
cryptography==43.0.3
cymem==2.0.8
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.16
dash-html-components==2.0.0
dash-labs==1.2.0
dash-local-react-components==1.3.0
dash-molstar==1.1.2
dash-table==5.0.0
dash-testing-stub==0.0.2
dataclass-wizard==0.22.3
deprecation==2.1.0
dill==0.3.9
EditorConfig==0.12.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
fire==0.7.0
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.5.0
gotrue==2.11.0
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.10
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==1.5.3
percy==2.0.2
plotly==5.22.0
pluggy==1.5.0
postgrest==0.18.0
preshed==3.0.9
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyOpenSSL==24.2.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.8
selenium==4.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tenacity==8.3.0
termcolor==2.5.0
text-unidecode==1.3
thinc==8.2.4
tomli==2.0.2
toolz==0.12.1
tqdm==4.66.4
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==1.26.20
urllib3-secure-extra==0.1.0
waitress==3.0.1
wasabi==1.1.3
weasel==0.4.1
websockets==13.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.19.1
